xdpi, ydpi int to flloat

+++ TeaLeaf/jni/platform/navigator.h
@@ -17,8 +17,8 @@
 
 typedef struct navigator_info_t {
     int density_dpi;
-    int xdpi;
-    int ydpi;
+    float xdpi;
+    float ydpi;
     char *android_version;
     char *language;
     char *country;
